Remove 'library-usage'
The last meaningful change to this tag was c4c48d49 (Change nonstatic
to library usage, 2013-07-10), but I'm not sure where that discussion
happened.  In any case, that commit changed some "must" wording to
"may" wording, which seems like it should move the label from required
to permitted.  However, a library-usage permission would also apply to
many other licenses (e.g. folks are free to link MIT-licensed work
from a proprietary program), and adding library-usage to almost all
the licenses seems like the wrong way to make this distinction [1].

The limitations that the LGPL and OSL place on disclose-source scoping
are already covered in the disclose-source description, so the
library-usage label doesn't seem to be adding anything meaningful.
The OSL gets at this distinction by tightly scoping derivative works
[2], and the LGPL talks about combined works as a special subset of
derivative works [3,4].  The MPL makes a similar distinction between
"Covered Software" and "Larger Work" [5], and the EPL makes a similar
distinction between "derivative works" and "the Program" [6].  Whether
the location of those distinctions, or the requirements placed on
combined works can be neatly summarized in a boolean label remains to
be seen, but we're pretty sure that library-usage is not that label
[7].

Subsequent commits may replace the caveat in the disclose-source
description with wording in the license description themselves or by
adding a new label that summarizes the issue.  Until then, the
disclose-source description more clearly covers the information that
library-usage was intended to convey, so this commit removes the
less-clear label to avoid redundancy.

[1]: https://github.com/github/choosealicense.com/pull/343#issuecomment-179532710
[2]: http://rosenlaw.com/OSL3.0-explained.htm#_Toc187293087
[3]: https://www.gnu.org/licenses/old-licenses/lgpl-2.1.html
[4]: https://www.gnu.org/licenses/lgpl.html
[5]: https://www.mozilla.org/en-US/MPL/2.0/
[6]: http://www.eclipse.org/legal/epl-v10.html
[7]: https://github.com/github/choosealicense.com/pull/343#issuecomment-179557468